+++ AccWideUILayoutSelection/LoadFunction.lua
@@ -523,13 +523,19 @@ function AccWideUIAceAddon:LoadUISettings(doNotLoadChatOrBagSettings)
 					end
 				end
 				
-				if (self:IsMidnight() ~= true) then
+				if (self:IsMainline() == true) then
+					if (self.db.profile.syncData.nameplates.special.NamePlateSize) then
+						C_NamePlate.SetNamePlateSize(self.db.profile.syncData.nameplates.special.NamePlateSize[1], self.db.profile.syncData.nameplates.special.NamePlateSize[2])
+					end
+				else
 					if (self.db.profile.syncData.nameplates.special.NamePlateEnemyClickThrough) then
 						C_NamePlate.SetNamePlateEnemyClickThrough(self.db.profile.syncData.nameplates.special.NamePlateEnemyClickThrough)
 					end
+					
 					if (self.db.profile.syncData.nameplates.special.NamePlateEnemyPreferredClickInsets) then
 						C_NamePlate.SetNamePlateEnemyPreferredClickInsets(self.db.profile.syncData.nameplates.special.NamePlateEnemyPreferredClickInsets[1], self.db.profile.syncData.nameplates.special.NamePlateEnemyPreferredClickInsets[2], self.db.profile.syncData.nameplates.special.NamePlateEnemyPreferredClickInsets[3], self.db.profile.syncData.nameplates.special.NamePlateEnemyPreferredClickInsets[4])
 					end
+					
 					if (self.db.profile.syncData.nameplates.special.NamePlateEnemySize) then
 						C_NamePlate.SetNamePlateEnemySize(self.db.profile.syncData.nameplates.special.NamePlateEnemySize[1], self.db.profile.syncData.nameplates.special.NamePlateEnemySize[2])
 					end
@@ -537,9 +543,11 @@ function AccWideUIAceAddon:LoadUISettings(doNotLoadChatOrBagSettings)
 					if (self.db.profile.syncData.nameplates.special.NamePlateFriendlyClickThrough) then
 						C_NamePlate.SetNamePlateFriendlyClickThrough(self.db.profile.syncData.nameplates.special.NamePlateFriendlyClickThrough)
 					end
+					
 					if (self.db.profile.syncData.nameplates.special.NamePlateFriendlyPreferredClickInsets) then
 						C_NamePlate.SetNamePlateFriendlyPreferredClickInsets(self.db.profile.syncData.nameplates.special.NamePlateFriendlyPreferredClickInsets[1], self.db.profile.syncData.nameplates.special.NamePlateFriendlyPreferredClickInsets[2], self.db.profile.syncData.nameplates.special.NamePlateFriendlyPreferredClickInsets[3], self.db.profile.syncData.nameplates.special.NamePlateFriendlyPreferredClickInsets[4])
 					end
+					
 					if (self.db.profile.syncData.nameplates.special.NamePlateFriendlySize) then
 						C_NamePlate.SetNamePlateFriendlySize(self.db.profile.syncData.nameplates.special.NamePlateFriendlySize[1], self.db.profile.syncData.nameplates.special.NamePlateFriendlySize[2])
 					end
@@ -547,9 +555,11 @@ function AccWideUIAceAddon:LoadUISettings(doNotLoadChatOrBagSettings)
 					if (self.db.profile.syncData.nameplates.special.NamePlateSelfClickThrough) then
 						C_NamePlate.SetNamePlateSelfClickThrough(self.db.profile.syncData.nameplates.special.NamePlateSelfClickThrough)
 					end
+					
 					if (self.db.profile.syncData.nameplates.special.NamePlateSelfPreferredClickInsets) then
 						C_NamePlate.SetNamePlateSelfPreferredClickInsets(self.db.profile.syncData.nameplates.special.NamePlateSelfPreferredClickInsets[1], self.db.profile.syncData.nameplates.special.NamePlateSelfPreferredClickInsets[2], self.db.profile.syncData.nameplates.special.NamePlateSelfPreferredClickInsets[3], self.db.profile.syncData.nameplates.special.NamePlateSelfPreferredClickInsets[4])
 					end
+					
 					if (self.db.profile.syncData.nameplates.special.NamePlateSelfSize) then
 						C_NamePlate.SetNamePlateSelfSize(self.db.profile.syncData.nameplates.special.NamePlateSelfSize[1], self.db.profile.syncData.nameplates.special.NamePlateSelfSize[2])
 					end
@@ -796,7 +806,7 @@ function AccWideUIAceAddon:LoadUISettings(doNotLoadChatOrBagSettings)
 			
 			
 			--  Midnight only settings
-			if (self:IsMidnight() == true) then
+			if (self:IsMainline() == true) then
 			
 				-- Use Damage Meter Setting
 				if (self.db.profile.syncToggles.damageMeter == true) then