+++ AccWideUILayoutSelection/Options.lua
@@ -960,22 +960,11 @@ function AccWideUIAceAddon:GenerateOptions()
 
 
 
-	-- Remove Sync options that are not applicable to various versions
-	--[[if (AccWideUIAceAddon:IsMidnight() == true) then
-		self.optionsData.args.settings.args.syncToggles.args.groupUnits.args.nameplates = nil
-	end]]
-
-	if (AccWideUIAceAddon:IsMidnight() ~= true) then
-		self.optionsData.args.settings.args.syncToggles.args.groupInterface.args.damageMeter = nil
-		self.optionsData.args.settings.args.syncToggles.args.groupInterface.args.externalDefensives = nil
-	end
-	
+	-- Remove Sync options that are not applicable to various versions	
 	if (AccWideUIAceAddon:IsMainline() == false) then
+		self.optionsData.args.settings.args.syncToggles.args.groupInterface.args.damageMeter = nil
 		self.optionsData.args.settings.args.editModeSettings = nil
 		self.optionsData.args.settings.args.headerDiv2 = nil
-	end
-
-	if (AccWideUIAceAddon:IsMainline() == false) then
 		self.optionsData.args.settings.args.syncToggles.args.groupInterface.args.cooldownViewer = nil
 		self.optionsData.args.settings.args.syncToggles.args.groupCombat.args.mouseoverCast = nil
 		self.optionsData.args.settings.args.syncToggles.args.groupCombat.args.empowerTap = nil